+++ lib/v2/wsj/radar.js
@@ -5,7 +5,7 @@ module.exports = {
             {
                 title: '新闻',
                 docs: 'https://docs.rsshub.app/traditional-media.html#hua-er-jie-ri-bao-the-wall-street-journal-wsj',
-                souce: '/',
+                source: '/',
                 target: '/wsj/zh-cn',
             },
         ],
@@ -13,7 +13,7 @@ module.exports = {
             {
                 title: '新闻',
                 docs: 'https://docs.rsshub.app/traditional-media.html#hua-er-jie-ri-bao-the-wall-street-journal-wsj',
-                souce: '/',
+                source: '/',
                 target: '/wsj/en-us',
             },
         ],
