+++ lib/v2/wsj/router.js
@@ -1,3 +1,3 @@
 module.exports = (router) => {
-    router.get('/:lang/:category?', require('./index'));
+    router.get('/:lang/:category?', require('./news'));
 };
